refactor: CarName 객체 생성하고 책임 이동
